Got ERC721 relayer working
Also reorganized code in different files so that it is easier to
understand and navigate.
